Refactors export client
